Format docs description

+++ app/src/main/java/ru/zakoulov/vkcupa/data/source/vk/models/VkDocument.kt
@@ -4,10 +4,10 @@ data class VkDocument(
     val id: Int,
     val ownerId: Int,
     val title: String,
-    val size: Int,
+    val size: Long,
     val ext: String,
     val url: String,
-    val date: Int,
+    val date: Long,
     val type: Int,
     val preview: VkPreview?,
     val tags: List<String>?
